add dryrun option

+++ scripts/01_upload_raw_dictionaries.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from classopt import classopt
+from classopt import classopt, config
 
 from aws_common import CredentialCache
 from fs_common import validate_file
@@ -16,6 +16,7 @@ class Opts:
     aws_region: str = "ap-northeast-1"
     s3_bucket: str = "sudachi"
     s3_prefix: str = "sudachidict-raw"
+    dryrun: bool = config(action="store_true")
 
 
 def validate_files(args: Opts) -> list[Path]:
@@ -34,6 +35,10 @@ def upload_files(client, args: Opts, files: list[Path]):
     bucket = client.Bucket(args.s3_bucket)
     for file in files:
         s3_key = f"{args.s3_prefix}/{args.version}/{file.name}"
+        if args.dryrun:
+            print("put", file, "to", s3_key)
+            continue
+
         with file.open('rb') as f:
             resp = bucket.put_object(
                 Body=f,
@@ -46,6 +51,10 @@ def upload_files(client, args: Opts, files: list[Path]):
 def regenerate_index(s3, args: Opts):
     listing = generate_raw_listing(s3, args.s3_bucket, args.s3_prefix)
     bucket = s3.Bucket(args.s3_bucket)
+
+    if args.dryrun:
+        print(listing)
+        return
 
     bucket.put_object(
         Body=listing.encode("utf-8"),
